Update gallerytesting.html
good, few more changes -
okay, and their is a lot of empty space on the side, is it possible to add a invisible border, and everything should be within it, and also as images increase, add a scrolling and also back to the top, right bottom up arrow button. Possible?