Update galleryembed.js
Implemented storing and retrieving from settings. For Viewer option, - bring in create a note, dump the embed data, then capture the note uuid and save it in settings, if setting has uuid, then open or redirect to it else create one and save it in settings and redirect to it!
Remove the time stamp from the note name - not required if the above program is implemented!