FINAL
Audit not possible and not happening for Viewer! - Its giving looped error!
Additionally when deleting the note of the Viewer is also creating issues + deleting the uuid details from the settings also causing issues!
Just once create the view and leave it alone, then every time you select the app options and viewer - it will take you to the note, hopefully